[models/ceremonies] extract current reputations to v1
